+++ StepperApp/Models/UserModel.swift
@@ -14,8 +14,8 @@ struct User: Codable {
     var isMan: Bool?
     var imageName: String
     var steps: Int = 0
-    var km: Int = 0
-    var miles: Int = 0
+    var km: Double = 0
+    var miles: Double = 0
 }
 
 struct Goal: Equatable {
